+++ disk_space/example/pubspec.yaml
@@ -21,8 +21,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
   flutter_lints: ^2.0.0
-
   integration_test:
     sdk: flutter
+
+flutter:
+  uses-material-design: true
