+++ disk_space_android/android/src/main/kotlin/br/com/suamusica/disk_space_android/MethodCallHandlerImpl.kt
@@ -33,8 +33,6 @@ internal class MethodCallHandlerImpl : MethodCallHandler{
         this.activity = activity
     }
 
-    
-
     private fun getFreeDiskSpace(): Double {
         val stat = StatFs(Environment.getExternalStorageDirectory().path)
